Move express app to server.js to allow it to be used by test suite
